Reconcile new Marker with existing gmaps.Marker.
This prevents the 'location' changes from being forgotten when rendering
Markers.

This change needs to be done for every property that may change from
*outside* of the map controllers!

(In this case, the Position of the Marker changes, but it's never
propagated to the map, unless the user explicitly subscribes to change
events on the Marker they've created).